Add handling of and warning information about the gyro range
The gyro range is a hidden parameter that causes endless headaches for
new users until they track down the calibration wiki online.  Attempt to
shortcut this process by giving a warning and pointing towards the wiki.
Additionally, display the current value before running calibration so you
don't waste time if you've forgotten to set it, and save the parameter to
both the launch file and the calibration yaml file.