Increase min max obstacle height for scan topic
Adjust the values so that it is possible to mount another laserscan source anywhere on the turtlebot.

In our setup we use, besides the kinect, a laserscanner which scans the field behind the turtlebot.
This scanner is placed above the kinect and publishes data with a higher value than the max_obstacle height,
in this case it is not possible to use the data for navigation and localization.